// GaroshaPrimoAngularUI/GaroshaPrimoAngularUIProject/src/app/Items/Models/item-model.model.ts
export class ItemModel {
    ItemId: number;
    ItemNameEnglish: string;
    ItemOtherName: string;
    ExpiryDate: Date;
    Stock: number;
    ItemSellingPrice: number;
    ProducerCompanyId: number;
    ItemBuyingDiscountPercentage: number;
    ItemBuyingPrice: number;
    TaxesPercentageOnBuying: number;
    TaxesValueOnBuying: number;
    TaxesPercentageOnSelling: number;
    TaxesValueOnSelling: number;
    ItemDescription: string;
    CreatedDate: Date;
    LastModified: Date;
    ItemStatus: string;
}