// GaroshaPrimoAngularUI/GaroshaPrimoAngularUIProject/src/app/Buying/Models/buying-transaction-item-display-model.model.ts
export class BuyingTransactionItemDisplayModel {
    ItemId: number;
    ItemNameEnglish: string;
    ItemExpiryDate: Date;
    BuyingTransactionItemQuantity: number;
    ItemSellingPrice: number;
    ItemBuyingDiscountPercentage: number;
    ItemBuyingPrice: number;
    ItemSubtotal: number;
    ItemBuyingTaxesPercentage: number;
    ItemBuyingTaxesValue: number;
    ItemTotal: number;
    Stock: number;
}